// SimpleCustomerService/SimpleCustomerService.Web/ClientApp/src/app/interfaces/ICustomer.ts
import { IOrder } from "./IOrder";
import { IState } from "./IState";

export interface ICustomer {
    id?: string;
    firstName: string;
    lastName: string;
    email: string;
    address: string;
    city: string;
    state?: IState;
    stateId?: number;
    zip: number;
    gender: string;
    orderCount?: number;
    orders?: IOrder[];
    orderTotal?: number;
}